Added script for measuring VM downtime with pings
This will continuously ping a host every 50ms for 1 min and return the highest consecutive amount of
time host was down.

To use the script run the command 'chmod u+x pingscript' and then './pingscript' in the directory of
the script. This script should be run on a seperate machine NOT the virtual machine that you are
transfering. Before using make sure to set the variable VIRTUAL_HOSTNAME to the correct hostname or
ip address.